flik文件

+++ lab4/flik/HorribleSteve.java
@@ -2,8 +2,8 @@ package flik;
 
 public class HorribleSteve {
     public static void main(String [] args) throws Exception {
-        int i = 0;
-        for (int j = 0; i < 500; ++i, ++j) {
+        long i = 0;
+        for (long j = 0; i < 500; ++i, ++j) {
             if (!Flik.isSameNumber(i, j)) {
                 throw new Exception(
                         String.format("i:%d not same as j:%d ??", i, j));
